nifi/script/requirements.txt


pymssql
selenium
protobuf==3.20.3
google-api-core>=2.0.0
google-cloud-core>=2.0.0
webdriver-manager